Jbullet physics engine added

+++ src/main/java/uib/bamboozle/Main.java
@@ -10,9 +10,9 @@ import java.io.BufferedReader;
 import java.io.InputStreamReader;
 
 public class Main {
-   public static int roll;
-   public static int pitch;
-   public static int yaw;
+   public static float roll;
+   public static float pitch;
+   public static float yaw;
    public static ReadFromGo reader;
 
    public static void main (String[] arg) {
